xtensa: fix example applications
Similar to i386, Xtensa also needs a global pointer instead of a
dedicated register. The Xtensa architecture always places literals
in front of the code. Depending on the amount of literals, the main
function is, therefore, located at different offsets from the file.
We add a simple jump to the beginning of the executable.

Signed-off-by: Chris Zankel <chris@zankel.net>